chore: 쿠폰 생성 api mutation 에러 핸들링 부분 mutate 함수에서 처리하도록 수정
