+++ src/hooks/mutations/useCreateCouponMutation.ts
@@ -1,21 +1,9 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { toast } from "react-toastify";
+import { useMutation } from "@tanstack/react-query";
 import { couponApi } from "@/apis/couponApi";
-import { QueryKey } from "@/constants/queryKey";
+import { CouponInfoType } from "@/types/entities/coupon";
 
 export default function useCreateCouponMutation() {
-  const queryClient = useQueryClient();
-
   return useMutation({
-    mutationFn: couponApi.createCoupon,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [QueryKey.couponList],
-      });
-      toast.success("쿠폰이 생성되었습니다.");
-    },
-    onError: (error: any) => {
-      toast.error(error.response.data.errorMessage);
-    },
+    mutationFn: (body: CouponInfoType) => couponApi.createCoupon(body),
   });
 }
